Input & Output
Integer Input
Python programming has several in-built functions to perform input and output tasks. Two commonly used functions for I/O (Input/Output) are print and raw_input(). Write a program to get an integer as input and print it. The raw_input() function reads formatted input from the keyboard. When the user enters an integer, it is stored in variable testInteger. NOTE: Bold letters or words are Input and rest are output statements.


Sample Input:



56



Sample Output:



56